requirements.txt
fastapi
uvicorn
pytesseract
Pillow
numpy
opencv-python-headless
google-generativeai
